Added the update function for profile + modify however needed
